yolo-detector/requirements.txt
Flask
gunicorn
ultralytics
Pillow